scrict handling of subfields
Default mode is strict mode, where invalid cases fail with an Exception.

For example adding a field with only empty subfields will result in an error.

To be more forgiving, set the Record.strict flag to False. This won't raise
errors, but tries to keep the record MARC conform by not adding subfields that
only contain empty values, for example.